+++ test/e2e/job.go
@@ -42,9 +42,9 @@ const (
 
 var _ = framework.KubeDescribe("Job", func() {
 	f := framework.NewDefaultFramework("job")
-	parallelism := 2
-	completions := 4
-	lotsOfFailures := 5 // more than completions
+	parallelism := int32(2)
+	completions := int32(4)
+	lotsOfFailures := int32(5) // more than completions
 
 	// Simplest case: all pods succeed promptly
 	It("should run a job to completion when tasks succeed", func() {
@@ -112,8 +112,8 @@ var _ = framework.KubeDescribe("Job", func() {
 	})
 
 	It("should scale a job up", func() {
-		startParallelism := 1
-		endParallelism := 2
+		startParallelism := int32(1)
+		endParallelism := int32(2)
 		By("Creating a job")
 		job := newTestJob("notTerminate", "scale-up", api.RestartPolicyNever, startParallelism, completions)
 		job, err := createJob(f.Client, f.Namespace.Name, job)
@@ -137,8 +137,8 @@ var _ = framework.KubeDescribe("Job", func() {
 	})
 
 	It("should scale a job down", func() {
-		startParallelism := 2
-		endParallelism := 1
+		startParallelism := int32(2)
+		endParallelism := int32(1)
 		By("Creating a job")
 		job := newTestJob("notTerminate", "scale-down", api.RestartPolicyNever, startParallelism, completions)
 		job, err := createJob(f.Client, f.Namespace.Name, job)
@@ -199,7 +199,7 @@ var _ = framework.KubeDescribe("Job", func() {
 })
 
 // newTestJob returns a job which does one of several testing behaviors.
-func newTestJob(behavior, name string, rPol api.RestartPolicy, parallelism, completions int) *batch.Job {
+func newTestJob(behavior, name string, rPol api.RestartPolicy, parallelism, completions int32) *batch.Job {
 	job := &batch.Job{
 		ObjectMeta: api.ObjectMeta{
 			Name: name,
@@ -270,7 +270,7 @@ func deleteJob(c *client.Client, ns, name string) error {
 }
 
 // Wait for all pods to become Running.  Only use when pods will run for a long time, or it will be racy.
-func waitForAllPodsRunning(c *client.Client, ns, jobName string, parallelism int) error {
+func waitForAllPodsRunning(c *client.Client, ns, jobName string, parallelism int32) error {
 	label := labels.SelectorFromSet(labels.Set(map[string]string{jobSelectorKey: jobName}))
 	return wait.Poll(framework.Poll, jobTimeout, func() (bool, error) {
 		options := api.ListOptions{LabelSelector: label}
@@ -278,7 +278,7 @@ func waitForAllPodsRunning(c *client.Client, ns, jobName string, parallelism int
 		if err != nil {
 			return false, err
 		}
-		count := 0
+		count := int32(0)
 		for _, p := range pods.Items {
 			if p.Status.Phase == api.PodRunning {
 				count++
@@ -289,7 +289,7 @@ func waitForAllPodsRunning(c *client.Client, ns, jobName string, parallelism int
 }
 
 // Wait for job to reach completions.
-func waitForJobFinish(c *client.Client, ns, jobName string, completions int) error {
+func waitForJobFinish(c *client.Client, ns, jobName string, completions int32) error {
 	return wait.Poll(framework.Poll, jobTimeout, func() (bool, error) {
 		curr, err := c.Extensions().Jobs(ns).Get(jobName)
 		if err != nil {
